feat: add dependencies and configuration files
